freeradius.py
This script does everything although at the moment it only tracks Authentications, info lines, and errors. It does not handle any accounting data as of right now but I have interest in adding that in the future. Also, currently this logwatch script prints everything regardless of what you have your logwatch level set to. This is also on my todo list.